renamed service package to downloadService
