+++ ase/src/main/java/com/ase/ase/AseCovidDashboardApplication.java
@@ -1,8 +1,8 @@
 package com.ase.ase;
 
 import com.ase.ase.dao.*;
-import com.ase.ase.services.DownloadPopulationData;
-import com.ase.ase.services.DownloadService;
+import com.ase.ase.downloadServices.DownloadPopulationData;
+import com.ase.ase.downloadServices.DownloadService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.CommandLineRunner;
