+++ ase/src/test/java/com/ase/ase/ExtractOverviewDataTests.java
@@ -1,7 +1,7 @@
 package com.ase.ase;
 
 import com.ase.ase.entities.Overview;
-import com.ase.ase.services.DownloadOverview;
+import com.ase.ase.downloadServices.DownloadOverview;
 import org.junit.jupiter.api.Test;
 
 import java.io.BufferedReader;
